requirements.txt
anyio==4.0.0
chardet==5.2.0
click==8.1.7
flasgger==0.9.7.1
Flask==2.2.3
Flask_Cors==4.0.0
Flask_Excel==0.0.7
pyexcel-xlsx==0.6.0
Flask_Migrate==4.0.5
flask_sqlalchemy==3.1.1
numpy==1.26.2
pydub==0.25.1
pyserial_asyncio==0.6
pyttsx3==2.90
PyYAML==6.0.1
SQLAlchemy==2.0.23
websockets==12.0
Werkzeug==2.2.3
PyMySQL==1.0.3
result==0.14.0
